Move playlist_mclick_actions to playlist_item_helpers namespace
