+++ foo_ui_columns/foo_ui_columns.vcxproj
@@ -445,6 +445,7 @@
     <ClCompile Include="config_host.cpp" />
     <ClCompile Include="list_view_panel.cpp" />
     <ClCompile Include="notification_area.cpp" />
+    <ClCompile Include="playlist_item_helpers.cpp" />
     <ClCompile Include="rebar_band.cpp" />
     <ClCompile Include="splitter_utils.cpp" />
     <ClCompile Include="tab_layout.cpp" />
@@ -510,6 +511,7 @@
     <ClInclude Include="filter_search_bar.h" />
     <ClInclude Include="list_view_panel.h" />
     <ClInclude Include="notification_area.h" />
+    <ClInclude Include="playlist_item_helpers.h" />
     <ClInclude Include="rebar_band.h" />
     <ClInclude Include="splitter_utils.h" />
     <ClInclude Include="tab_layout.h" />
