+++ foo_ui_columns/main_window.cpp
@@ -38,38 +38,6 @@ bool remember_window_pos()
     return config_object::g_get_data_bool_simple(standard_config_objects::bool_remember_window_positions, false);
 }
 
-void action_remove_track(bool on_item, unsigned idx)
-{
-    if (on_item) {
-        static_api_ptr_t<playlist_manager> api;
-        api->activeplaylist_undo_backup();
-        api->activeplaylist_remove_items(pfc::bit_array_one(idx));
-    }
-}
-
-void action_add_to_queue(bool on_item, unsigned idx)
-{
-    if (on_item) {
-        static_api_ptr_t<playlist_manager> api;
-        unsigned active = api->get_active_playlist();
-        if (active != -1)
-            api->queue_add_item_playlist(active, idx);
-    }
-}
-
-void action_none(bool on_on_item, unsigned idx) {}
-
-pma playlist_mclick_actions::g_pma_actions[] = {
-    {"(None)", 0, action_none},
-    {"Remove track from playlist", 1, action_remove_track},
-    {"Add to playback queue", 2, action_add_to_queue},
-};
-
-unsigned playlist_mclick_actions::get_count()
-{
-    return tabsize(g_pma_actions);
-}
-
 const TCHAR* main_window_class_name = _T("{E7076D1C-A7BF-4f39-B771-BCBE88F2A2A8}");
 
 const wchar_t* unsupported_os_message
@@ -333,16 +301,6 @@ void cui::MainWindow::resize_child_windows()
             }
         }
     }
-}
-
-unsigned playlist_mclick_actions::id_to_idx(unsigned id)
-{
-    unsigned count = tabsize(g_pma_actions);
-    for (unsigned n = 0; n < count; n++) {
-        if (g_pma_actions[n].id == id)
-            return n;
-    }
-    return 0;
 }
 
 bool process_keydown(UINT msg, LPARAM lp, WPARAM wp, bool playlist, bool keyb)
