+++ foo_ui_columns/ng_playlist/ng_playlist.cpp
@@ -7,6 +7,7 @@
 #include "ng_playlist.h"
 #include "ng_playlist_groups.h"
 #include "../config_columns_v2.h"
+#include "../playlist_item_helpers.h"
 
 namespace artwork_panel {
 // extern cfg_string cfg_front;
@@ -934,7 +935,7 @@ const style_data_t& ng_playlist_view_t::get_style_data(t_size index)
 }
 bool ng_playlist_view_t::notify_on_middleclick(bool on_item, t_size index)
 {
-    return playlist_mclick_actions::run(cfg_playlist_middle_action, on_item, index);
+    return cui::playlist_item_helpers::mclick_action::run(cfg_playlist_middle_action, on_item, index);
 }
 bool ng_playlist_view_t::notify_on_doubleleftclick_nowhere()
 {
